* Updated all other robots to work with base class * Added take profit level based on stop loss level (inpTakeProfitRiskRewardRatio)

+++ MetaTraderStrategies/jerrold.cpp
@@ -11,9 +11,9 @@
 
 // Money management / risk parameters
 input double   _inpLots = 1;                // Number of lots to trade
-input double   _inpStopLossPips = 30;       // Initial stop loss in pips
+input int      _inpStopLossPips = 30;       // Initial stop loss in pips
 input bool     _inpUseTakeProfit = true;    // Whether to use a take profit order or not
-input double   _inpTakeProfitPips = 40;     // Take profit level in pips
+input int      _inpTakeProfitPips = 40;     // Take profit level in pips
 input int      _inpTrailingStopPips = 30;   // Trailing stop in pips (0 to not use a trailing stop)
 
 // Pin Bar parameters
